debugging

+++ pigolf/app.py
@@ -176,7 +176,7 @@ class App(tk.Frame):
         self.height = 768
         self.resolution = f"{self.width}x{self.height}"
         self.framerate = 25
-        self.delay = 1000/self.framerate
+        self.delay = int(1000/self.framerate)
 
         self.queue = queue.Queue()
